Move the metadata-previews ProgressBar component to the components package and replace the old one
