+++ packages/components/tests/ProgressBarTest.js
@@ -1,0 +1,33 @@
+import React from "react";
+import renderer from "react-test-renderer";
+
+import ProgressBar from "../ProgressBar";
+
+describe( "The Progressbar component", () => {
+	it( "without any props matches the snapshot", () => {
+		const component = renderer.create(
+			<ProgressBar />
+		);
+
+		const tree = component.toJSON();
+		expect( tree ).toMatchSnapshot();
+	} );
+
+	test( "with progress matches the snapshot", () => {
+		const component = renderer.create(
+			<ProgressBar max={ 100 } value={ 50 } />
+		);
+
+		const tree = component.toJSON();
+		expect( tree ).toMatchSnapshot();
+	} );
+
+	test( "with custom color matches the snapshot", () => {
+		const component = renderer.create(
+			<ProgressBar progressColor="#000" max={ 100 } value={ 50 } />
+		);
+
+		const tree = component.toJSON();
+		expect( tree ).toMatchSnapshot();
+	} );
+});
