+++ packages/search-metadata-previews/src/index.js
@@ -1,5 +1,4 @@
 export { default as constants } from "./SnippetPreview/constants"
-export { default as ProgressBar } from "./SnippetPreview/ProgressBar";
 export * from "./SnippetPreview/SnippetPreview";
 export { default as HelpTextWrapper} from "./SnippetPreview/HelpTextWrapper";
 export { default as FixedWidthContainer} from "./SnippetPreview/FixedWidthContainer"
